Introduce shorter numeric type aliases and resolve implicit conversion issues

+++ include/cecxx/benchmark/detail/cec2017/evaluate.hpp
@@ -8,7 +8,7 @@
 
 namespace cecxx::benchmark::detail::cec2017 {
 
-auto dispatch_problem(problem_context_view ctx, const int fn, const auto& input) {
+auto dispatch_problem(problem_context_view ctx, const u8 fn, const auto& input) {
   switch (fn) {
   case 1:
     return bent_cigar(input, ctx);
@@ -75,14 +75,14 @@ auto dispatch_problem(problem_context_view ctx, const int fn, const auto& input)
   throw std::runtime_error{"Unknown CEC problem"};
 }
 
-auto evaluate(problem_context_view ctx, const int fn_num, const auto& input) {
+auto evaluate(problem_context_view ctx, const u8 fn_num, const auto& input) {
   const auto ncol = input.size();
   const auto nrow = input.at(0).size();
   if (not cec2017::VALID_DIMENSIONS.contains(nrow)) {
     throw std::runtime_error{"Invalid problem dimension."};
   }
 
-  auto output = std::vector<double>(ncol);
+  auto output = std::vector<f64>(ncol);
   for (auto col{0u}; col < output.size(); ++col) {
     output[col] = dispatch_problem(ctx, fn_num, input.at(col)) + cec2017::CEC2017_OFFSET.at(fn_num);
   }
