+++ src/table_io.hpp
@@ -15,18 +15,18 @@ struct FILE_deleter_t {
 };
 
 struct table_size_t {
-  int size{};
-  int scaler{};
+  u64 size{};
+  i64 scaler{};
   bool scaler_applied{false};
 };
 
-auto ROT_TABLE_FILENAME(const std::filesystem::path& datadir, const int dim, const int fn) -> std::string;
-auto SHUFFLE_TABLE_FILENAME(const std::filesystem::path& datadir, const int dim, const int fn) -> std::string;
-auto SHIFT_TABLE_FILENAME(const std::filesystem::path& datadir, const int, const int fn) -> std::string;
+auto ROT_TABLE_FILENAME(const std::filesystem::path& datadir, const u8 dim, const u8 fn) -> std::string;
+auto SHUFFLE_TABLE_FILENAME(const std::filesystem::path& datadir, const u8 dim, const u8 fn) -> std::string;
+auto SHIFT_TABLE_FILENAME(const std::filesystem::path& datadir, const u8, const u8 fn) -> std::string;
 
-auto ROT_TABLE_SIZE(const int dim, const int fn) -> table_size_t;
-auto SHIFT_TABLE_SIZE(const int dim, const int fn) -> table_size_t;
-auto SHUFFLE_TABLE_SIZE(const int dim, const int fn) -> table_size_t;
+auto ROT_TABLE_SIZE(const u8 dim, const u8 fn) -> table_size_t;
+auto SHIFT_TABLE_SIZE(const u8 dim, const u8 fn) -> table_size_t;
+auto SHUFFLE_TABLE_SIZE(const u8 dim, const u8 fn) -> table_size_t;
 
 template <table_type_t Table, typename... Args>
 constexpr auto get_table_name(const std::filesystem::path& datapath, Args... args) {
@@ -42,7 +42,7 @@ constexpr auto get_table_name(const std::filesystem::path& datapath, Args... arg
   throw std::runtime_error("Unknown table type.");
 }
 
-template <table_type_t Table, typename ValueType> constexpr auto get_table_size(const int dim, const int fn) {
+template <table_type_t Table, typename ValueType> constexpr auto get_table_size(const u8 dim, const u8 fn) {
   switch (Table) {
   case table_type_t::rotate:
     return ROT_TABLE_SIZE(dim, fn);
@@ -55,13 +55,13 @@ template <table_type_t Table, typename ValueType> constexpr auto get_table_size(
   throw std::runtime_error("Unknown table type.");
 }
 
-template <table_type_t Table, numeric T> auto scan_table(auto* fs, const int dim, const int fn) -> std::vector<T> {
+template <table_type_t Table, numeric T> auto scan_table(auto* fs, const u8 dim, const u8 fn) -> std::vector<T> {
   const auto sz_info = get_table_size<Table, T>(dim, fn);
   auto table = std::vector<T>(sz_info.size);
 
   const auto scan_scaled_shift_table = [&]() {
-    for (auto i = 0; i < sz_info.scaler - 1; ++i) {
-      for (auto j = 0; j < dim; ++j) {
+    for (auto i = 0u; i < sz_info.scaler - 1; ++i) {
+      for (auto j = 0u; j < dim; ++j) {
         if (std::fscanf(fs, number_formatter<T>().data(), &table[i * dim + j]) == -1) {
           throw std::runtime_error{"Failed to parse table data from file."};
         }
@@ -75,7 +75,7 @@ template <table_type_t Table, numeric T> auto scan_table(auto* fs, const int dim
     return table;
   }
 
-  for (auto i = 0; i < sz_info.size; ++i) {
+  for (auto i = 0u; i < sz_info.size; ++i) {
     if (std::fscanf(fs, number_formatter<T>().data(), &table[i]) == -1) {
       throw std::runtime_error{"Failed to parse table data from file."};
     }
@@ -84,7 +84,7 @@ template <table_type_t Table, numeric T> auto scan_table(auto* fs, const int dim
 }
 
 template <table_type_t Table, numeric T>
-auto load_table_data(const std::filesystem::path& datapath, const int dim, const int fn) {
+auto load_table_data(const std::filesystem::path& datapath, const u8 dim, const u8 fn) {
   const auto filename = get_table_name<Table>(datapath, dim, fn);
   auto fs_ptr = std::unique_ptr<std::FILE, FILE_deleter_t>{std::fopen(filename.data(), "r")};
   if (not fs_ptr) {
